Remove references to $MANGOH_ROOT from cdef files
The goal was to remove explicit paths where possible.  In the case of
the submodules, this helps to make them more useful as a standalone
repository in case a use isn't interested in cloning the whole mangOH
repository and all of the submodules.

Also updated the developerDay2016 tutorial app to use AVC 2.0 so that
it will work with Legato versions 17.05 and newer.