iot-slot: Fix bug in GPIO requests
GPIO requests were passing pointers to stack allocated strings under the
assumption that the label string contents would be copied to another
storage location during the request.  This assumption was incorrect.
The string pointer is just copied, so the pointer becomes unsafe to
follow once the requesting function returns or re-uses the string.