Introduce a unified mangOH.sdef
Introduce a single mangOH.sdef and remove the board/module specific
SDEFs.  The content of the single SDEF change depending on the build
parameters.  The "targets" directory was introduced based on the way
Legato handles different modules.  The location of build artifacts has
now changed.  Everything produced by the Makefile is under "build/".