Re-enable battery gauge on red for wp76/77
The battery gauge was disabled on wp76/77 because of a lockup in the
power supply kernel subsystem when the battery gauge driver was
unloaded.  This has been fixed in release 8 and thus the battery gauge
driver and battery service can be enabled.  The battery gauge and
battery service still aren't usable on wp75/85 because of a different
kernel bug.