Simplify Makefile
The mangOH build is failing in the developer VM based on Ubuntu 18.04
because the wrong toolchain is being selected for the red_wp85 build.  I
think the root cause may have to do with the .phony target that sets up
the Make variables not being run multiple times.  Rather than spending a
lot of time on this, I have just simplified the Makefile so that the
problem goes away.