Introduced CONFIG_IIO_CHAN_TYPE_GASRESISTANCE
Previous commit 354e27d introduced a dependency upon IIO_GASRESISTANCE
being defined as a member of enum iio_chan_type.  The enum is part of
the kernel headers and not part of the out-of-tree IIO code in the
mangOH repository.  The code now checks that
CONFIG_IIO_CHAN_TYPE_GASRESISTANCE is defined before trying to reference
IIO_GASRESISTANCE.