Remove kernel mod dependency bug workaround
Prior to Legato 18.07.0, there was bug (LE-10550) in Legato which caused
the supervisor to get stuck in an infinite loop if two different mdefs
included in a system depended on the same mdef.  As a workaround the
dependency of bmp280.mdef on iio-kfifo-buf.mdef was commented out.  Now
that this bug is resolved, the dependency can be restored.