Introduce rtc-pcf85063 and rtc_sync kernel modules
rtc-pcf85063 is an I2C RTC chip that exists on the mangOH Yellow. This
driver was copied from the mainline Linux kernel. rtc_sync is a kernel
module that sets the system time from rtc1 when the module is loaded and
then periodically syncs the system time to the RTC if the system time is
changed.