Remove gpsSensor from mangOH.sdef
Remove gpsSensor app from the default mangOH.sdef because it creates a
lot of noise in the logs when it's not able to get a GPS fix. Also, the
location values it produces are intentionally slightly wrong to simulate
movement. This may confuse users who would expect the GPS location to be
accurate.