mangoh red: add bmi160 interrupt provision support
The bmi160 interrupt lines are connected to CF3 GPIO 38 and 36 by
default.  There is a hardware provision to connect the interrupt pins to
GPIO expander GPIO 11 and 12.  This commit introduces a define
BMI160_INT_TO_GPIO_EXPANDER which if defined assumes the interrupt pins
are connected to the GPIO expander GPIOs instead of the CF3 GPIOs.