fix error check in probe() for mangoh yellow
Fix a copy/paste error in the mangOH yellow module. Previously, the IMU
variable was being checked after calling i2c_new_device to create the
magnetometer. The impact of this bug is that a failed probe of the
magnetometer would not have been detected during probe of the mangOH
yellow driver.