Work around spiService start-up race
The spiService was failing to start up because the spisvc
kernel module was slow to create its device file and the
Supervisor was trying to add that file to the spiService's
working directory before the file was created.

Worked around this by making the spisvc load at Legato
start-up instead of at app start-up.

Unfortunately, this requires forking a bunch of stuff
from the Legato framework into the mangOH project, so we
can change it slightly.