UI tweaks re. enable/disable out-of-box exp
Pulled instant gratification features enable/disable code out of
the Octave screen so other screens could access it too.

Improved the wording on the Octave screen.

Added an option on the main menu to enable/disable the out-of-box
experience features.

Made the LED screen detect that the out-of-box experience is
disabled, explain the consequences, and provide the option of
re-enabling it.