Make more friendly when don't have Octave apps
Print a helpful error if OCTAVE=1 and OCTAVE_ROOT not set.

Default OCTAVE=0 (so build succeeds) when OCTAVE_ROOT is not set
and OCTAVE is not defined. This will minimize interruption of
work for people who don't have access to the Octave apps or
the sources needed to build them (which is most people right now).