bluetoothSensorTag: support CC1350 SensorTag (#61)
Update to support the CC1350 SensorTag in addition to the already
supported CC2650.  The CC1350 doesn't have a green LED, so the
toggle_leds.sh script has been updated to also control the buzzer.  This
makes it easier for an observer to distinguish the red LED blinking from
the advertising pattern (also red LED blinking).