opt300x: fix race condition
When performing a read of the light sensor, an end of conversion
interrupt is received.  The interrupt signals to another thread that a
reading is ready using a wait_queue.  Immediately after that, it checks
ok_to_ignore_lock to see if it should unlock the mutex.  The problem is
that there is a chance that the newly awakened other thread has updated
ok_to_ignore_lock before the interrupt thread tests the value.  If this
occurs, the mutex will be unlocked twice.