remove default.sdef copied from Legato
The only difference between the mangOH version of the default.sdef and
the original copied from Legato is that the mangOH version removed the
smsInboxService app. This app was originally removed to save space, but
it was later found that removing the devMode app was a far more
effective way to save space.

Resolves: #75