yellow module: write EEPROM if it's empty
This mangOH Yellow kernel module update causes the initialization code
to assume that the device is the latest revision of the mangOH yellow if
the EEPROM is blank and write the EEPROM with a string indicating that
it's a mangOH yellow.  A flag is added to the EEPROM to indicate that
this is not the result of a successful factory test process.