Allow for light sensor without irq on yellow
There is no low power MCU within WP7610 and WP7611 modules.  On mangOH
Yellow, the interrupt pin of the opt3002 light sensor is connected to
one of the GPIOs that is expected to be provided by the low power MCU.
Since this GPIO won't be available in these modules, we don't set the
irq member of the i2c_board_info struct in the case that the GPIO
request fails.  It won't be possible to receive high/low threshold
interrupts from the light sensor on these modules.